Move metabase/core/utils/arrays to metabase/lib/arrays, and move metabase/core/utils/errors to metabase/lib/errors
