+++ frontend/src/metabase/account/profile/components/UserProfileForm/UserProfileForm.tsx
@@ -7,7 +7,7 @@ import FormInput from "metabase/core/components/FormInput";
 import FormSelect from "metabase/core/components/FormSelect";
 import FormSubmitButton from "metabase/core/components/FormSubmitButton";
 import FormErrorMessage from "metabase/core/components/FormErrorMessage";
-import * as Errors from "metabase/core/utils/errors";
+import * as Errors from "metabase/lib/errors";
 import type { LocaleData, User } from "metabase-types/api";
 import type { UserProfileData } from "../../types";
