+++ frontend/src/metabase/actions/containers/ActionCreator/FormCreator/utils.ts
@@ -1,5 +1,5 @@
 import _ from "underscore";
-import { moveElement } from "metabase/core/utils/arrays";
+import { moveElement } from "metabase/lib/arrays";
 import type { FieldSettingsMap, InputSettingType } from "metabase-types/api";
 
 const inputTypeMap: Record<InputSettingType, string> = {
