+++ frontend/src/metabase/admin/databases/components/DatabaseEditApp/Sidebar/ModelActionsSection/ModelActionsSection.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import { t } from "ttag";
 
 import Toggle from "metabase/core/components/Toggle";
-import { getResponseErrorMessage } from "metabase/core/utils/errors";
+import { getResponseErrorMessage } from "metabase/lib/errors";
 
 import {
   ToggleContainer,
