+++ frontend/src/metabase/admin/settings/slack/components/SlackForm/SlackForm.tsx
@@ -5,7 +5,7 @@ import { Form, FormProvider } from "metabase/forms";
 import FormInput from "metabase/core/components/FormInput";
 import FormSubmitButton from "metabase/core/components/FormSubmitButton";
 import FormErrorMessage from "metabase/core/components/FormErrorMessage";
-import * as Errors from "metabase/core/utils/errors";
+import * as Errors from "metabase/lib/errors";
 import type { SlackSettings } from "metabase-types/api";
 import { SlackFormMessage } from "./SlackForm.styled";
