+++ frontend/src/metabase/auth/components/LoginForm/LoginForm.tsx
@@ -6,7 +6,7 @@ import FormCheckBox from "metabase/core/components/FormCheckBox";
 import FormErrorMessage from "metabase/core/components/FormErrorMessage";
 import FormInput from "metabase/core/components/FormInput";
 import FormSubmitButton from "metabase/core/components/FormSubmitButton";
-import * as Errors from "metabase/core/utils/errors";
+import * as Errors from "metabase/lib/errors";
 import type { LoginData } from "../../types";
 
 const LOGIN_SCHEMA = Yup.object().shape({
