+++ frontend/src/metabase/collections/components/CreateCollectionForm/CreateCollectionForm.tsx
@@ -13,7 +13,7 @@ import FormTextArea from "metabase/core/components/FormTextArea";
 import FormSubmitButton from "metabase/core/components/FormSubmitButton";
 import FormErrorMessage from "metabase/core/components/FormErrorMessage";
 
-import * as Errors from "metabase/core/utils/errors";
+import * as Errors from "metabase/lib/errors";
 
 import { color } from "metabase/lib/colors";
