--- frontend/src/metabase/core/utils/arrays/index.ts
@@ -1,1 +1,0 @@
-export * from "./arrays";
