+++ frontend/src/metabase/databases/utils/schema.ts
@@ -1,6 +1,6 @@
 import * as Yup from "yup";
 import type { TestContext } from "yup";
-import * as Errors from "metabase/core/utils/errors";
+import * as Errors from "metabase/lib/errors";
 import type { DatabaseData, Engine, EngineField } from "metabase-types/api";
 import { ADVANCED_FIELDS, FIELD_OVERRIDES } from "../constants";
