+++ frontend/src/metabase/forms/hooks/use-form-submit/use-form-submit.ts
@@ -1,6 +1,6 @@
 import { useCallback, useState } from "react";
 import type { FormikHelpers } from "formik";
-import { getResponseErrorMessage } from "metabase/core/utils/errors";
+import { getResponseErrorMessage } from "metabase/lib/errors";
 import type { FormState } from "../../contexts";
 import type { FormError } from "./types";
