+++ frontend/src/metabase/lib/errors/index.ts
@@ -1,4 +1,5 @@
 export * from "./messages";
 export * from "./validation";
+export * from "./server-error-types";
 
 export type { GenericErrorResponse } from "./types";
