+++ frontend/src/metabase/lib/errors/server-error-types.ts
