+++ frontend/src/metabase/lib/errors/types.ts
