+++ frontend/src/metabase/query_builder/components/NativeQueryEditor/NativeQueryEditorPrompt/NativeQueryEditorPrompt.tsx
@@ -8,7 +8,7 @@ import LoadingSpinner from "metabase/components/LoadingSpinner";
 import type { DatabaseId } from "metabase-types/api";
 import { MetabotApi } from "metabase/services";
 import Tooltip from "metabase/core/components/Tooltip";
-import { getResponseErrorMessage } from "metabase/core/utils/errors";
+import { getResponseErrorMessage } from "metabase/lib/errors";
 import type { Deferred } from "metabase/lib/promise";
 import { defer } from "metabase/lib/promise";
 import {
