+++ frontend/src/metabase/setup/components/NewsletterForm/NewsletterForm.tsx
@@ -4,7 +4,7 @@ import * as Yup from "yup";
 import { useSelector } from "metabase/lib/redux";
 import { FormProvider } from "metabase/forms";
 import FormSubmitButton from "metabase/core/components/FormSubmitButton";
-import * as Errors from "metabase/core/utils/errors";
+import * as Errors from "metabase/lib/errors";
 import type { SubscribeInfo } from "metabase-types/store";
 import { subscribeToNewsletter } from "../../utils";
 import { getUserEmail } from "../../selectors";
